requirements.txt
flask
requests
openai
beautifulsoup4
faiss-cpu
numpy
